Added functionality to parse only TIFF & GeoTIFF headers, added overload for Loadxxx() functions, more type changes.

+++ GeoTIFF-Parser/GeoTIFF_Parser_Globals.h
@@ -47,16 +47,16 @@ public:
 struct GeoTIFFDetails
 {
 public:
-	unsigned short int rasterSpace; //0: undefined, 1: PixelIsArea, 2:PixelIsPoint, 32767: User Defined.
-	unsigned short int modelType; //0: Undefined or Unknown, 1: 2D Projected CRS, 2: Geographic 2D CRS, 3: Cartesian 3D CRS, 32767: User Defined.
+	unsigned int rasterSpace; //0: undefined, 1: PixelIsArea, 2:PixelIsPoint, 32767: User Defined.
+	unsigned int modelType; //0: Undefined or Unknown, 1: 2D Projected CRS, 2: Geographic 2D CRS, 3: Cartesian 3D CRS, 32767: User Defined.
 	RasterToModelTransformationMethod transformationMethod = RasterToModelTransformationMethod::unknown;
 	double tiePoints[2][3]; //Only used if transformationMethod is tieAndScale.
 	double pixelScale[3]; //XYZ scale. For XY, positive scale indicate increase in XY coord as raster space UV increase, negatives denote an inverse relation. Only used if transformationMethod is tieAndScale.
 	double modelTransformationMatrix[4][4]; // A 4x4 transformation matrix used to transform from raster to model space, when transformationMethod is matrix.
 
-	unsigned short int projectedCRS; //Ranges 1-1023 reserved, 1024-32766 EPSG Projected CRS Codes, 32767 is User Defined, 32768-65535 are private.
-	unsigned short int geodeticCRS; //Ranges 1-1023 reserved, 1024-32766 EPSG Geographic 2D or Geocentric CRS, 32767 User Defined, 32768-65535 private.
-	unsigned short int verticalCRS; //Ranges 1-1023 reserved, 1024-32766 EPSG Geographic 2D or Geocentric CRS, 32767 User Defined, 32768-65535 private.
+	unsigned int projectedCRS; //Ranges 1-1023 reserved, 1024-32766 EPSG Projected CRS Codes, 32767 is User Defined, 32768-65535 are private.
+	unsigned int geodeticCRS; //Ranges 1-1023 reserved, 1024-32766 EPSG Geographic 2D or Geocentric CRS, 32767 User Defined, 32768-65535 private.
+	unsigned int verticalCRS; //Ranges 1-1023 reserved, 1024-32766 EPSG Geographic 2D or Geocentric CRS, 32767 User Defined, 32768-65535 private.
 
 	std::string geotiffCitation = "";
 	std::string geodeticCRSCitation = "";
@@ -80,8 +80,8 @@ public:
 
 struct Tag
 {
-	unsigned short int tagID;
-	unsigned short int fieldTypeID;
+	unsigned int tagID;
+	unsigned int fieldTypeID;
 	unsigned long int count;
 	unsigned long int offsetValue;
 
@@ -91,10 +91,10 @@ struct Tag
 
 struct GeoKey
 {
-	unsigned short int keyID;
-	unsigned short int tiffTagLocation; //if ==0, then offsetVale contains the value (of type short int) and count is implied to be = 1.
-	unsigned short int count;
-	unsigned short int offsetValue;
+	unsigned int keyID;
+	unsigned int tiffTagLocation; //if ==0, then offsetVale contains the value (of type short int) and count is implied to be = 1.
+	unsigned int count;
+	unsigned int offsetValue;
 };
 
 struct Type
@@ -112,12 +112,12 @@ public:
 };
 
 
+//TODO move the variables into files where they are filled (all in GeoTIFF_Parser.cpp, I think). Add functions to that file that return pointers to const versions of them.
 //Variables
 extern bool isBigEndian;
 extern std::ifstream stream;
 extern TIFFDetails tiffDetails;
 extern GeoTIFFDetails geoDetails;
-
 
 //Functions
 long int BytesToInt32(const char bytes[4]);
